requirements.txt
python-dotenv
langchain-groq
langchain-ollama
